Add property to sorter
